requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-socketio==0.0.10
Flask==3.0.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openai==1.10.0
paramiko==3.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyNaCl==1.5.0
python-engineio==4.8.2
python-socketio==5.11.0
redis==5.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
